Reference App/server/requirements.txt
connexion==2.6.0
swagger-ui-bundle==0.0.6
python_dateutil==2.8.1
setuptools>=45.1.0
pyjwt==1.7.1
gunicorn>=20.0.0
flask-cors==3.0.8
openpyxl>=3.0.3
pandas~=0.25.1
